Tests, black, rm personal test file
